コンポーネント分割/autofocus実装
