+++ frontend/src/functions/getContestId.ts
@@ -1,0 +1,4 @@
+export function getContestId() {
+  const splitPath = window.location.pathname.split('/');
+  return splitPath.slice(-1)[0];
+}
